optimizer tidy up and row major fix

+++ msg/include/Message.h
@@ -5,7 +5,7 @@
 
 template <typename Type>
 struct Message : public Type {
-	std::uint32_t timestamp;
+	std::uint64_t timestamp;
 	std::string src;
 };
